# Conflicts:
#	groups/bdl/bdlb/bdlb_guidutil.h
#	groups/bdl/bdlb/doc/bdlb.txt
#	groups/bdl/bdlc/doc/bdlc.txt
#	groups/bdl/bdlma/doc/bdlma.txt
#	groups/bdl/bdls/doc/bdls.txt
#	groups/bdl/bdls/package/bdls.opts
#	groups/bdl/bdlt/bdlt_datetimeinterval.h
#	groups/bdl/bdlt/doc/bdlt.txt
#	groups/bdl/doc/bdl.txt
#	groups/bsl/bslh/doc/bslh.txt
#	groups/bsl/bslmf/doc/bslmf.txt
#	groups/bsl/bsls/doc/bsls.txt
#	groups/bsl/bslstl/doc/bslstl.txt
#	groups/bsl/bslx/doc/bslx.txt

+++ groups/bdl/bdlma/bdlma_multipoolallocator.cpp
@@ -11,9 +11,9 @@ BSLS_IDENT_RCSID(bdlma_multipoolallocator_cpp,"$Id$ $CSID$")
 namespace BloombergLP {
 namespace bdlma {
 
-                     // ------------------------
-                     // class MultipoolAllocator
-                     // ------------------------
+                         // ------------------------
+                         // class MultipoolAllocator
+                         // ------------------------
 
 // CREATORS
 MultipoolAllocator::~MultipoolAllocator()
@@ -28,7 +28,7 @@ void *MultipoolAllocator::allocate(size_type size)
         return 0;                                                     // RETURN
     }
 
-    return d_multipool.allocate(size);
+    return d_multipool.allocate(static_cast<int>(size));
 }
 
 void MultipoolAllocator::deallocate(void *address)
@@ -46,14 +46,15 @@ void MultipoolAllocator::reserveCapacity(size_type size, size_type numObjects)
         return;                                                       // RETURN
     }
 
-    d_multipool.reserveCapacity(size, numObjects);
+    d_multipool.reserveCapacity(static_cast<int>(size),
+                                static_cast<int>(numObjects));
 }
 
 }  // close package namespace
 }  // close enterprise namespace
 
 // ----------------------------------------------------------------------------
-// Copyright 2012 Bloomberg Finance L.P.
+// Copyright 2015 Bloomberg Finance L.P.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
